Move pattern cve_2019_15548_2 into cve_2019_15548
